Introduce AbstractParallelCollector
