--- src/main/java/com/pivovarit/collectors/CollectorUtils.java
@@ -1,35 +1,0 @@
-package com.pivovarit.collectors;
-
-import java.util.Collection;
-import java.util.concurrent.CompletableFuture;
-import java.util.function.BiFunction;
-import java.util.function.BinaryOperator;
-import java.util.function.Supplier;
-
-final class CollectorUtils {
-
-    private CollectorUtils() {
-    }
-
-    static <T1, R1 extends Collection<T1>> BinaryOperator<CompletableFuture<R1>> mergingPartialResults() {
-        return (f1, f2) -> f1.thenCombine(f2, (left, right) -> {
-            left.addAll(right);
-            return left;
-        });
-    }
-
-    static <T1, R1 extends Collection<T1>> BiFunction<CompletableFuture<R1>, CompletableFuture<T1>, CompletableFuture<R1>> accumulatingResults() {
-        return (list, object) -> list.thenCombine(object, (left, right) -> {
-            left.add(right);
-            return left;
-        });
-    }
-
-    static <T1> T1 supplyWithResources(Supplier<T1> supplier, Runnable action) {
-        try {
-            return supplier.get();
-        } finally {
-            action.run();
-        }
-    }
-}
