+++ src/main/java/com/pivovarit/collectors/UnboundedParallelCollector.java
@@ -1,6 +1,5 @@
 package com.pivovarit.collectors;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.EnumSet;
 import java.util.List;
@@ -11,25 +10,17 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.Executor;
 import java.util.function.BiConsumer;
-import java.util.function.BinaryOperator;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.stream.Collector;
 
-import static com.pivovarit.collectors.CollectorUtils.accumulatingResults;
-import static com.pivovarit.collectors.CollectorUtils.mergingPartialResults;
-import static com.pivovarit.collectors.CollectorUtils.supplyWithResources;
 import static java.util.concurrent.CompletableFuture.completedFuture;
-import static java.util.concurrent.CompletableFuture.supplyAsync;
-import static java.util.concurrent.Executors.newSingleThreadExecutor;
 
 /**
  * @author Grzegorz Piwowarek
  */
 final class UnboundedParallelCollector<T, R, C extends Collection<R>>
-  implements Collector<T, List<CompletableFuture<R>>, CompletableFuture<C>>, AutoCloseable {
-
-    private volatile boolean isFailed = false;
+  extends AbstractParallelCollector<T, R, C>
+  implements AutoCloseable {
 
     private final ParallelDispatcher<R> dispatcher;
 
@@ -55,24 +46,11 @@ final class UnboundedParallelCollector<T, R, C extends Collection<R>>
     }
 
     @Override
-    public Supplier<List<CompletableFuture<R>>> supplier() {
-        return ArrayList::new;
-    }
-
-    @Override
-    public BinaryOperator<List<CompletableFuture<R>>> combiner() {
-        return (left, right) -> {
-            left.addAll(right);
-            return left;
-        };
-    }
-
-    @Override
     public BiConsumer<List<CompletableFuture<R>>, T> accumulator() {
         return (acc, e) -> {
             CompletableFuture<R> future = new CompletableFuture<>();
             dispatcher.addPending(future);
-            dispatcher.addTask(() -> isFailed ? null : operation.apply(e));
+            dispatcher.addTask(() -> dispatcher.isMarkedFailed() ? null : operation.apply(e));
             acc.add(future);
         };
     }
@@ -81,7 +59,7 @@ final class UnboundedParallelCollector<T, R, C extends Collection<R>>
     public Function<List<CompletableFuture<R>>, CompletableFuture<C>> finisher() {
         if (dispatcher.isNotEmpty()) {
             dispatcher.start();
-            return foldLeftFutures().andThen(f -> supplyWithResources(() -> f, dispatcher::close));
+            return foldLeftFutures(collectionFactory).andThen(f -> supplyWithResources(() -> f, dispatcher::close));
         } else {
             return supplyWithResources(() -> (__) -> completedFuture(collectionFactory.get()), dispatcher::close);
         }
@@ -103,7 +81,7 @@ final class UnboundedParallelCollector<T, R, C extends Collection<R>>
             while ((task = tasks.poll()) != null && !Thread.currentThread().isInterrupted()) {
 
                 try {
-                    if (isFailed) {
+                    if (dispatcher.isMarkedFailed()) {
                         dispatcher.cancelAll();
                         break;
                     }
@@ -124,19 +102,12 @@ final class UnboundedParallelCollector<T, R, C extends Collection<R>>
                   next.complete(r);
               } else {
                   next.completeExceptionally(throwable);
-                  isFailed = true;
+                  dispatcher.markFailed();
               }
           });
     }
 
     private void closeAndCompleteRemaining(Exception e) {
         dispatcher.closeExceptionally(e);
-    }
-
-    private Function<List<CompletableFuture<R>>, CompletableFuture<C>> foldLeftFutures() {
-        return futures -> futures.stream()
-          .reduce(completedFuture(collectionFactory.get()),
-            accumulatingResults(),
-            mergingPartialResults());
     }
 }
